DDR-990..1004: lwIP hammer, OPEN-1 decision, multi-arch assessment, and four defects found by their own gates (#14)
Merged on three CONFIRMED greens on one tip, d0a85b5, per §INV.15 — and all
three were INDEPENDENT runs rather than re-attempts of one:

  push             33272639659  completed, 0 of 15 jobs failed
  pull_request     33272641911  completed, 0 of 15 jobs failed
  workflow_dispatch 33274262876  completed, 0 of 15 jobs failed

Release candidate verified on the shipping kernel bb9c6187a30bb0dd
(1,098,122 B against the 1,572,864 B gate): smoke-iso-x86 PASS on both the
BIOS and UEFI arms of one ISO, and smoke-iso-userspace PASS — the ISO boots a
live OS: SFS root + PRISM + AETHER agent + write/read/delete round-trip.
ISO 1f3ca48c73c51c5b, 52,805,632 B.

What this branch settles:

DDR-990 — the two-CPU connect/close hammer, mutation-checked in both
directions, proving the lwIP g_net_lock fix rather than failing to disprove it.

DDR-999/1000 — OPEN-1 decided in writing. Routes 2 and 3 CLOSED (route 2 at 95%
power, 60/60 on one kernel hash against a threshold set before the run); route 1
remains open, scope narrowed to a CI-only hang with no artefact. And an honest
multi-arch parity assessment: aarch64/riscv64 stay boot-only for 1.0.

DDR-1001..1004 — four defects, each found by a gate rather than guessed at:
  * the wait4 ring walk took no lock (fixed, bounded + locked);
  * GLOBAL_FORBIDDEN had been the EMPTY STRING for four commits because a
    comment sat inside a printf continuation — every gate in the repo ran with
    no global safety net, and smoke-selftest case 5 is what caught it;
  * the PTE_SW_SHARED audit: invariant HOLDS across all nine vmm_map_in sites
    and across fork, but ptnode_in_use underflows on every fork (recorded
    unfixed — no gate observes it, so §NON-NEGOTIABLE 3 forbids the fix);
  * the rq-3 resched proof demanded an IPI the kernel only promises when an
    idle CPU is visible, so a correct system failed its own test.

DDR-997 §13 — and one defect in a fix from this same branch: raising the
resizeall per-round waits made a retrying run exceed the guest cap, so it was
SIGTERM'd mid-arm and an arm that never executed printed FAIL in the same words
a real failure uses. The injection budget is now derived from that cap and
"never ran" is a distinct verdict.

Not claimed: OPEN-1 route 1, OPEN-12 and OPEN-13 stay open. v1.0.0 stays
untagged — that hold is the operator's to lift.

+++ kernel/apic/smp.c
@@ -181,7 +181,7 @@ void smp_start_aps(void) {
 
     /* Stage the trampoline at 0x8000 (low conventional RAM, reserved by
      * construction — the kernel lives at 0x400000 since DDR-733). */
-    uint64_t tlen = (uint64_t)(ap_tramp_end - ap_tramp_start);
+    uint64_t tlen = (uint64_t)((uintptr_t)ap_tramp_end - (uintptr_t)ap_tramp_start);
     volatile uint8_t *tramp = (volatile uint8_t *)(uintptr_t)TRAMP_PHYS;
     for (uint64_t i = 0; i < tlen; i++)
         tramp[i] = ap_tramp_start[i];
